feat: move TransactionItem class to common package
